pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelmaker"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
mcp = ["mcp[cli]>=1.2.0"]

[project.scripts]
pixelmaker = "pixelmaker.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pixelmaker*"]
